db configurations

+++ src/main/java/kr/di/uoa/gr/jedaiwebapp/models/MethodConfigurationRepository.java
@@ -3,7 +3,7 @@ package kr.di.uoa.gr.jedaiwebapp.models;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface MethodConfigurationRepository extends CrudRepository<MethodConfiguration, Double>{
+public interface MethodConfigurationRepository extends CrudRepository<MethodConfiguration, Integer>{
 	
 	MethodConfiguration findById(int id);
